feature: Dog detector feature added to Jupyter Notebook
The dog detector feature was added using the VGG16 model after comparing to the VGG19 and Alexnet models. These models were all pretrained on the ImageNet dataset. The objective of using these models is to identify from the 1000 classes in the ImageNet dataset in a given image, but only to give a result if a dog class is detected (which ranges between 151 - 268 on the ImageNet class list).